pip-requirements.txt
tornado==4.3
websocket-client==0.34.0
coverage
supervisor
